FeatureクラスのファイルからEntityクラスを別ファイルの切り出し
