+++ Macho/MachoFramework/Sources/MachoView/Views/DiaryCreation/Entity/Goal.swift
@@ -1,0 +1,18 @@
+//
+//  MachoFramework
+//
+//  Goal.swift
+//
+//  Created by stotic-dev on 2025/03/09
+//  Copyright © Macho All rights reserved.
+//
+
+import Foundation
+
+struct Goal: Equatable, Identifiable {
+    
+    let id: UUID
+    var trainingType: TrainingTypeData
+    var numberOfSets: Int
+    var setCount: Int
+}
